requirements.txt
pytest==7.4.0
requests==2.31.0
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-xdist==3.3.1
allure-python-commons~=2.13.2